backend/api_auto_generated/requirements.txt
aiofiles>=24.1.0
aniso8601>=9.0.1
annotated-types>=0.7.0
anyio>=3.7.1
async-exit-stack>=1.0.1
async-generator>=1.10
bcrypt>=4.0.1
certifi>=2024.7.4
cffi>=1.16.0
chardet>=5.2.0
charset-normalizer>=3.3.2
click>=8.1.7
cryptography>=42.0.8
dnspython>=2.6.1
email_validator>=2.2.0
fastapi>=0.111.0
fastapi-cli>=0.0.4
graphene>=3.3
graphql-core>=3.2.3
graphql-relay>=3.2.0
h11>=0.14.0
httpcore>=1.0.5
httptools>=0.6.1
httpx>=0.27.0
idna>=3.7
iniconfig>=2.0.0
itsdangerous>=2.2.0
Jinja2>=3.1.4
markdown-it-py>=3.0.0
MarkupSafe>=2.1.5
mdurl>=0.1.2
orjson>=3.10.6
packaging>=24.1
passlib>=1.7.4
pluggy>=1.5.0
promise>=2.3
pycparser>=2.22
pydantic>=2.8.2
pydantic_core>=2.20.1
Pygments>=2.18.0
PyJWT>=2.8.0
pytest>=8.2.2
python-dotenv>=1.0.1
python-multipart>=0.0.9
PyYAML>=6.0.1
requests>=2.32.3
rich>=13.7.1
Rx>=3.2.0
shellingham>=1.5.4
six>=1.16.0
sniffio>=1.3.1
starlette>=0.37.2
typer>=0.12.3
typing_extensions>=4.12.2
ujson>=5.10.0
urllib3>=2.2.2
uvicorn>=0.30.1
uvloop>=0.19.0
watchfiles>=0.22.0
watchgod>=0.8.2
websockets>=12.0
pytest >= 8.0.0
sqlalchemy >= 2.0.0

